% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggutils.R
\name{easy_change_legend}
\alias{easy_change_legend}
\title{Easily modify legends}
\usage{
easy_change_legend(what = c("position", "direction", "justification"), to)
}
\arguments{
\item{what}{legend component to modify
(\code{"position"}, \code{"direction"}, or \code{"justification"})}

\item{to}{to what to set the legend component should be changed}
}
\value{
a \code{\link[ggplot2]{theme}} object
}
\description{
Change legend position, direction, or justification.
}
\details{
Due to limitations of \code{ggplot2} this will apply to all legends at once
}
\examples{

library(ggplot2)

# Move legends to bottom
ggplot(mtcars, aes(wt, mpg, colour = cyl, size = hp)) +
  geom_point() + easy_move_legend("bottom")

# Make legends horizontal
ggplot(mtcars, aes(wt, mpg, colour = cyl, size = hp)) +
  geom_point() + easy_rotate_legend("horizontal")

# Justify legends to the bottom and justify to the right
ggplot(mtcars, aes(wt, mpg, colour = cyl, size = hp)) +
  geom_point() +
  easy_move_legend("bottom") +
  easy_adjust_legend("right")
}
\author{
Jonathan Carroll
}
