% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_date.R
\name{fuzzy_date_inner_join}
\alias{fuzzy_date_inner_join}
\alias{fuzzy_date_left_join}
\alias{fuzzy_date_right_join}
\alias{fuzzy_date_semi_join}
\alias{fuzzy_date_anti_join}
\alias{fuzzy_date_full_join}
\title{fuzzy_date_inner_join}
\usage{
fuzzy_date_inner_join(df1, df2, by, time_buffer, ...)

fuzzy_date_left_join(df1, df2, by, time_buffer, ...)

fuzzy_date_right_join(df1, df2, by, time_buffer, ...)

fuzzy_date_semi_join(df1, df2, by, time_buffer, ...)

fuzzy_date_anti_join(df1, df2, by, time_buffer, ...)

fuzzy_date_full_join(df1, df2, by, time_buffer, ...)
}
\arguments{
\item{df1}{first candidate to be joined}

\item{df2}{second candidate to be joined}

\item{by}{Date class joining column}

\item{time_buffer}{numeric}

\item{...}{PARAM_DESCRIPTION}

\item{time_buffer_type}{time buffer i.e. months/ years/ minutes}
}
\value{
inner joined df

left joined df

right joined df

semi joined df

anti joined df

full joined df
}
\description{
join dates by a time buffer

join dates by a time buffer

join dates by a time buffer

join dates by a time buffer

join dates by a time buffer

join dates by a time buffer
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  library(tibble)
  library(lubridate)
  library(dplyr)
  df_1 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:27:10", 34.8014080, 103.8499800,
    "2018-09-30 04:55:51", 43.3367432, 44.158934,
    "2018-02-28 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))


  df_2 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:57:10", 34.8014080, 103.8999800,
    "2018-09-30 04:55:51", 43.3367432, 48.158934,
    "2018-02-27 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))

  fuzzy_date_inner_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 1
  )

  fuzzy_date_inner_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 3
  )
}
}
\dontrun{
if (interactive()) {
  library(tibble)
  library(lubridate)
  df_1 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:27:10", 34.8014080, 103.8499800,
    "2018-09-30 04:55:51", 43.3367432, 44.158934,
    "2018-02-28 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))


  df_2 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:57:10", 34.8014080, 103.8999800,
    "2018-09-30 04:55:51", 43.3367432, 48.158934,
    "2018-02-27 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))

  fuzzy_date_left_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 1
  )

  fuzzy_date_left_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 3
  )
}
}
\dontrun{
if (interactive()) {
  library(tibble)
  library(lubridate)
  df_1 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:27:10", 34.8014080, 103.8499800,
    "2018-09-30 04:55:51", 43.3367432, 44.158934,
    "2018-02-28 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))


  df_2 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:57:10", 34.8014080, 103.8999800,
    "2018-09-30 04:55:51", 43.3367432, 48.158934,
    "2018-02-27 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))

  fuzzy_date_right_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 1
  )

  fuzzy_date_right_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 3
  )
}
}
\dontrun{
if (interactive()) {
  library(tibble)
  library(lubridate)
  df_1 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:27:10", 34.8014080, 103.8499800,
    "2018-09-30 04:55:51", 43.3367432, 44.158934,
    "2018-02-28 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))


  df_2 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:57:10", 34.8014080, 103.8999800,
    "2018-09-30 04:55:51", 43.3367432, 48.158934,
    "2018-02-27 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))

  fuzzy_date_semi_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 1
  )

  fuzzy_date_semi_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 3
  )
}
}
\dontrun{
if (interactive()) {
  library(tibble)
  library(lubridate)
  df_1 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:27:10", 34.8014080, 103.8499800,
    "2018-09-30 04:55:51", 43.3367432, 44.158934,
    "2018-02-28 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))


  df_2 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:57:10", 34.8014080, 103.8999800,
    "2018-09-30 04:55:51", 43.3367432, 48.158934,
    "2018-02-27 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))

  fuzzy_date_anti_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 1
  )

  fuzzy_date_anti_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 3
  )
}
}
\dontrun{
if (interactive()) {
  library(tibble)
  library(lubridate)
  df_1 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:27:10", 34.8014080, 103.8499800,
    "2018-09-30 04:55:51", 43.3367432, 44.158934,
    "2018-02-28 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))


  df_2 <- tribble(
    ~datetime, ~latitude, ~longitude,
    "2018-10-01 08:57:10", 34.8014080, 103.8999800,
    "2018-09-30 04:55:51", 43.3367432, 48.158934,
    "2018-02-27 17:03:27", 37.0399910, 115.6672080
  ) \%>\%
    mutate(datetime = ymd_hms(datetime))

  fuzzy_date_full_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 1
  )

  fuzzy_date_full_join(df_1, df_2,
    by = "datetime",
    time_buffer_type = "days",
    time_buffer = 3
  )
}
}
}
\seealso{
\code{\link[lubridate]{period}}, \code{\link[lubridate]{character(0)}}
\code{\link[fuzzyjoin]{difference_join}}

\code{\link[lubridate]{period}}, \code{\link[lubridate]{character(0)}}
\code{\link[fuzzyjoin]{difference_join}}

\code{\link[lubridate]{period}}, \code{\link[lubridate]{character(0)}}
\code{\link[fuzzyjoin]{difference_join}}

\code{\link[lubridate]{period}}, \code{\link[lubridate]{character(0)}}
\code{\link[fuzzyjoin]{difference_join}}

\code{\link[lubridate]{period}}, \code{\link[lubridate]{character(0)}}
\code{\link[fuzzyjoin]{difference_join}}

\code{\link[lubridate]{period}}, \code{\link[lubridate]{character(0)}}
\code{\link[fuzzyjoin]{difference_join}}
}
