% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibily.R
\name{possibly2}
\alias{possibly2}
\title{possibly2}
\usage{
possibly2(.f, otherwise = NULL)
}
\arguments{
\item{.f}{generic function}

\item{otherwise}{either static out or a function, Default: NULL}
}
\value{
possibly class function
}
\description{
a purrr::possibly wrapper to generate dynamic default when failing
}
\details{
it consistnetly works well with purrr::insistently
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
