% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggutils.R
\name{scale_y_continuous_euro}
\alias{scale_y_continuous_euro}
\alias{scale_x_continuous_euro}
\title{scale_y_continuous_euro}
\usage{
scale_y_continuous_euro(..., prefix = "", suffix = "")

scale_x_continuous_euro(..., prefix = "", suffix = "")
}
\arguments{
\item{...}{inherited}

\item{prefix}{inherited, Default: ''}

\item{suffix}{inherited, Default: ''}
}
\value{
scale_y_continuous_euro with € suffix

scale_x_continuous_euro with € suffix
}
\description{
scales y var and adds € suffix

scales x var and adds € suffix
}
\examples{
\dontrun{
if (interactive()) {
  txhousing \%>\%
    filter(city == "El Paso") \%>\%
    group_by(year) \%>\%
    summarise(
      mean_volume_by_year = mean(volume)
    ) \%>\%
    ggplot(aes(x = year, y = mean_volume_by_year / 10^6)) +
    geom_line() +
    scale_y_continuous_euro(suffix = "M ") +
    labs(
      x = "",
      y = "mean volume in Mil (€) "
    ) +
    theme_ipsum_rc()
}
}
}
\seealso{
\code{\link[ggplot2]{scale_continuous}}
\code{\link[scales]{dollar_format}}

\code{\link[ggplot2]{scale_continuous}}
\code{\link[scales]{dollar_format}}
}
