% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_utils.R
\name{url_exists}
\alias{url_exists}
\title{url_exists}
\usage{
url_exists(url)
}
\arguments{
\item{url}{string url}
}
\value{
bool
}
\description{
it checks if a url exists or not
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{http_status}}, \code{\link[httr]{GET}}
}
