% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_utils.R
\name{url_exists2}
\alias{url_exists2}
\title{url_exists2}
\usage{
url_exists2(x, non_2xx_return_value = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{str url}

\item{non_2xx_return_value}{exclude 2xx errors, Default: FALSE}

\item{quiet}{non verbose out, Default: FALSE}

\item{...}{further}
}
\value{
bool
}
\description{
it checks if an url exists or not
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[httr]{HEAD}}, \code{\link[httr]{GET}}, \code{\link[httr]{status_code}}
}
